Client Argument Processing and Cleanup
Add argument processing to the basic client. Accepts the IP of the
server (-i) and the port of the server (-p). IP defaults to 0.0.0.0 and
port defaults to 5001

Move socket connection code out of the main and into its own function.

Client only starts and sends a debut text message.